Island Bounding Box Generation Done
Islands now will have their location tied to an island based on its ID,
and the town will be randomly placed on the island it corresponds with.
All this information as well will be saved and loaded properly.
Tweaked town radius a bit to be able to be reached by the player still